First Set Computation

+++ src/parser/set.c
@@ -4,21 +4,21 @@
 token_set *createTokenSet()
 {
     token_set *ts = (token_set *)malloc(sizeof(token_set));
-    ts->set = 0;
+    ts->set = 0LL;
     return ts;
 }
 
 void addToken(token_set *ts, TOKEN tok)
 {
-    ts->set |= ((long long)1 << tok);
+    ts->set |= (1LL << tok);
 }
 
 void removeToken(token_set *ts, TOKEN tok)
 {
-    ts->set &= ~((long long)1 << tok);
+    ts->set &= ~(1LL << tok);
 }
 
 int isMember(token_set *ts, TOKEN tok)
 {
-    return (ts->set & ((long long)1 << tok)) != (long long)0;
+    return (ts->set & (1LL << tok)) != 0LL;
 }
